ClarifEye.AI/ColorDetector/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
matplotlib
pillow
tqdm
torchsummary
fastapi 
uvicorn
python-multipart